upgrade PyQt5 and numpy

+++ t_view/model/RoiData.py
@@ -121,7 +121,7 @@ class RoiDataManager():
                                    0.75 * (img_dimension[0]) - 1,
                                    (2 * ind + 1) * part_height - 1,
                                    (2 * ind + 2) * part_height - 1])
-                limits = np.round(limits).astype(np.int)
+                limits = np.round(limits).astype(np.int32)
                 rois.append(Roi(limits))
 
             self._add(img_dimension, rois)
@@ -156,7 +156,7 @@ def get_roi_sum(img, roi):
     """
     roi = validate_roi(roi)
     roi_img = img[int(roi.y_min):int(roi.y_max) + 1, int(roi.x_min):int(roi.x_max) + 1]
-    return np.sum(roi_img, 0) / np.float(np.size(roi_img, 0))
+    return np.sum(roi_img, 0) / np.float32(np.size(roi_img, 0))
 
 
 def get_roi_max(img, roi):
